Remove nosetests
