+++ ckanext/datagovuk/tests/test_helpers.py
@@ -51,6 +51,8 @@ class TestHelpers:
             "education": "Education",
             "health": "Health",
             "transport": "Transport",
+            "digital-services-performance": "Digital services performance",
+            "government-reference-data": "Government reference data"
         }
         assert h.themes() == themes_dict
